+++ Cargo.toml
@@ -23,3 +23,8 @@ nu-ansi-term = "0.49.0"
 proc-macro2 = { version = "1.0.66", features = ["span-locations"] }
 syn = { version = "2.0.28", features = ["full"] }
 thiserror = "1.0.44"
+
+[dev-dependencies]
+anyhow = "1.0.98"
+seq-macro = "0.3.6"
+tempfile = "3.20.0"
