+++ tests/golden_1.rs
@@ -1,0 +1,6 @@
+fn foo_used() {}
+
+#[allow(dead_code)]
+fn main() {
+    foo_used();
+}
