+++ tests/golden_2.rs
@@ -1,0 +1,10 @@
+#[derive(Default)]
+struct AliveStruct {
+    x: i32,
+}
+
+#[allow(dead_code)]
+fn main() {
+    let s = AliveStruct::default();
+    println!("{}", s.x);
+}
