+++ tests/golden_3.rs
@@ -1,0 +1,8 @@
+mod foo {
+    pub fn foo_used() {}
+}
+
+#[allow(dead_code)]
+fn main() {
+    foo::foo_used();
+}
