+++ tests/golden_4.rs
@@ -1,0 +1,4 @@
+mod mymod {}
+
+#[allow(dead_code)]
+fn main() {}
