+++ tests/input_1.rs
@@ -1,0 +1,8 @@
+fn foo_used() {}
+
+fn foo_unused() {}
+
+#[allow(dead_code)]
+fn main() {
+    foo_used();
+}
